requirements.txt
requests
google-genai
python-dotenv
feedparser
pywebpush
cryptography
